+++ android/app/src/main/java/app/plusone/guestlist/MainActivity.java
@@ -1,0 +1,5 @@
+package app.plusone.guestlist;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}